CI-465 main updated without insights update

+++ src/Sfa.Das.ApprenticeshipInfoService.Infrastructure/Elasticsearch/ElasticsearchCustomClient.cs
@@ -48,7 +48,7 @@ namespace Sfa.Das.ApprenticeshipInfoService.Infrastructure.Elasticsearch
                 Identifier = identifier,
                 ReturnCode = result.ApiCall?.HttpStatusCode,
                 Successful = result.ApiCall?.Success,
-                SearchTime = result.Took,
+                SearchTime = (int)result.Took,
                 RequestTime = Math.Round(clientRequestTime.TotalMilliseconds, 2),
                 MaxScore = result.MaxScore,
                 HitCount = result.Hits?.Count(),
